prototype/requirements.txt
python-dotenv
openai
gradio
ollama
beautifulsoup4
requests
python-docx
pdfplumber
